Set up Carto user name and API key with environment

All functions need a Carto user name and API key.

setup_key()

Value

setup status message

Details

run file.edit("~/.Renviron") to edit the environment variable file

add two lines

carto_acc = "your user name"

carto_api_key = "your api key"

run setup_key().

Note if you want to remove the key and deleted the lines from ~/.Renviron, the key could still exist in environment. Restart R session to make sure it was removed.

For adding key or changing key value, edit and run setup_key() is enough.